+++ test/page-ex.test.js
@@ -396,7 +396,6 @@ describe('Page extensions', function() {
       page.end = function() {
         expect(page.app.convert.getCall(0).args[0]).to.equal('Hello');
         expect(page.app.convert.getCall(0).args[1]).to.equal('md');
-        expect(page.app.convert.getCall(0).args[2]).to.deep.equal({});
         expect(page.app.render.getCall(0).args[0]).to.equal('index');
         expect(page.app.render.getCall(0).args[1]).to.deep.equal({
           content: '<p>Hello</p>',
@@ -412,30 +411,37 @@ describe('Page extensions', function() {
       page.compile('Hello', 'md', 'index');
     }); // should convert markup, render layout and write page
     
-    /*
-    it('should convert markup and render', function(done) {
+    it('should convert markup, render layout and write page with options', function(done) {
       var app = new function(){};
       app.convert = sinon.stub().returns('<p>Hello</p>');
+      app.render = sinon.stub().yieldsAsync(null, '<p>Hello</p>');
       
       var page = new Page();
       setPrototypeOf(page, Object.create(pagex, {
           app: { configurable: true, enumerable: true, writable: true, value: app }
         }));
         
-      page.render = function(layout, options) {
+      page.write = sinon.spy();
+      page.end = function() {
         expect(page.app.convert.getCall(0).args[0]).to.equal('Hello');
         expect(page.app.convert.getCall(0).args[1]).to.equal('md');
-        expect(layout).to.equal('greet');
-        expect(options).to.deep.equal({ content: '<p>Hello</p>' });
+        expect(page.app.convert.getCall(0).args[2].pedantic).to.be.true;
+        expect(page.app.render.getCall(0).args[0]).to.equal('index');
+        expect(page.app.render.getCall(0).args[1]).to.deep.equal({
+          content: '<p>Hello</p>',
+          name: 'Tobi',
+          _locals: undefined,
+          pedantic: true
+        });
+        expect(page.write.getCall(0).args[0]).to.equal('<p>Hello</p>');
         done();
-      }
+      };
       page.next = function() {
         done('Page#next should not be called');
       };
       
-      page.compile('Hello', 'md', 'greet');
-    }); // should convert markup and render
-    */
+      page.compile('Hello', 'md', 'index', { name: 'Tobi', pedantic: true });
+    }); // should convert markup, render layout and write page with options
     
   }); // #compile